requirements_host.txt
opencv-python>=3.4.3,<4
PyYAML>=3.13,<4
scipy>=1.1,<2
numpy>=1.15,<2

# Tests
mock>=2,<3
pytest>=3.9.2,<4